fix warning (-Wuseless-cast)

+++ single_include/nlohmann/json.hpp
@@ -8277,7 +8277,7 @@ class binary_reader
             case 0x9B: // array (eight-byte uint64_t for n follow)
             {
                 std::uint64_t len{};
-                return get_number(input_format_t::cbor, len) && get_cbor_array(static_cast<std::size_t>(len), tag_handler);
+                return get_number(input_format_t::cbor, len) && get_cbor_array(len, tag_handler);
             }
 
             case 0x9F: // array (indefinite length)
@@ -8331,7 +8331,7 @@ class binary_reader
             case 0xBB: // map (eight-byte uint64_t for n follow)
             {
                 std::uint64_t len{};
-                return get_number(input_format_t::cbor, len) && get_cbor_object(static_cast<std::size_t>(len), tag_handler);
+                return get_number(input_format_t::cbor, len) && get_cbor_object(len, tag_handler);
             }
 
             case 0xBF: // map (indefinite length)
